fix: LikeController error! #929

+++ protected/modules_core/like/controllers/LikeController.php
@@ -71,7 +71,9 @@ class LikeController extends ContentAddonController
         if (!Yii::app()->user->isGuest) {
             $attributes = array('object_model' => $this->contentModel, 'object_id' => $this->contentId, 'created_by' => Yii::app()->user->id);
             $like = Like::model()->findByAttributes($attributes);
-            $like->delete();
+            if (!empty($like)) {
+                $like->delete();
+            }
         }
 
         $this->actionShowLikes();
